Extract VQ overlay from MissionMapWidget into VanquishMapOverlayWidget
